obexd: Fix the types of session->size and session->transferred
These were represented as ssize_t which represent the maximum size
of the addressable memory, so they are 32bits in 32bit machines. We
were expecting them to be 64bits.

+++ obexd/client/session.h
@@ -48,8 +48,8 @@ struct session_data {
 	char *buffer;
 	size_t buffer_len;
 	int filled;
-	ssize_t size;
-	ssize_t transferred;
+	gint64 size;
+	gint64 transferred;
 	gchar *filename;
 	gchar *agent_name;
 	gchar *agent_path;
